v1.6.4
Allowed `.markdown-preview-view` to have an `inherit` font-size, which in concert with the 3rd party plug-in Query Control, allows Linked Mentions to be rendered with decent enough font sizes. Beforehand, the filenames were smaller than the body text being previewed. Now the body text is smaller. There is still too much spacing for the body text, but that's a problem I can't figure out.